requirements.txt
pandas
requests
python-dotenv
matplotlib
flask
plotly
ecbdata
kaleido
pandasdmx
gunicorn